Add helpers folder with common functions
