+++ src/helpers/helpers.core.js
@@ -1,0 +1,65 @@
+import {isObject} from 'chart.js/helpers';
+
+// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/flat
+export function flatten(input) {
+  const stack = [...input];
+  const res = [];
+  while (stack.length) {
+    // pop value from stack
+    const next = stack.pop();
+    if (Array.isArray(next)) {
+      // push back array items, won't modify the original input
+      stack.push(...next);
+    } else {
+      res.push(next);
+    }
+  }
+  // reverse to restore input order
+  return res.reverse();
+}
+
+export function index(values, key) {
+  let n = values.length;
+  let i;
+
+  if (!n) {
+    return key;
+  }
+
+  const obj = isObject(values[0]);
+  key = obj ? key : 'v';
+
+  for (i = 0, n = values.length; i < n; ++i) {
+    if (obj) {
+      values[i]._idx = i;
+    } else {
+      values[i] = {v: values[i], _idx: i};
+    }
+  }
+  return key;
+}
+
+export function sort(values, key) {
+  if (key) {
+    values.sort((a, b) => +b[key] - +a[key]);
+  } else {
+    values.sort((a, b) => +b - +a);
+  }
+}
+
+export function sum(values, key) {
+  let s, i, n;
+
+  for (s = 0, i = 0, n = values.length; i < n; ++i) {
+    s += key ? +values[i][key] : +values[i];
+  }
+
+  return s;
+}
+
+export function requireVersion(min, ver) {
+  const parts = ver.split('.');
+  if (!min.split('.').reduce((a, c, i) => a && c <= parts[i], true)) {
+    throw new Error(`Chart.js v${ver} is not supported. v${min} or newer is required.`);
+  }
+}
