+++ src/index.js
@@ -3,5 +3,3 @@ import TreemapController from './controller';
 import TreemapElement from './element';
 
 Chart.register(TreemapController, TreemapElement);
-
-export * from './utils';
